reorganise cmake stuff
