+++ core/lz4/src/lz4.h
