+++ core/lzo/CMakeLists.txt
@@ -5,6 +5,7 @@
 #---Declare ZipLZO sources as part of libCore-------------------------------
 set(sources ${CMAKE_CURRENT_SOURCE_DIR}/src/ZipLZO.c)
 
+include_directories(${LZO_INCLUDE_DIR})
 
 ROOT_OBJECT_LIBRARY(Lzo ${sources})
